# src/main/resources/static/docs/1.0.0/swagger/todo-1.0.0-swagger.json
{
  "openapi" : "3.0.1",
  "info" : {
    "description" : "API de lista de afazeres\n",
    "version" : "1.0.0",
    "title" : "Todo list API",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "name" : "Lucas Bovolini",
      "email" : "lbovolini94@gmail.com"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "servers" : [ {
    "description" : "SwaggerHub API Auto Mocking",
    "url" : "https://virtserver.swaggerhub.com/lbovolini/Todo/1.0.0"
  }, {
    "description" : "localhost",
    "url" : "http://localhost:8080/todo"
  } ],
...
  "paths" : {
    "/login" : {
      "post" : {
        "tags" : [ "Authentication API" ],
        "summary" : "Autentica o usuário no sistema",
        "operationId" : "login",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UserCredentials"
              },
              "example" : {
                "username" : "lucas",
                "password" : "pass11word"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Operação realizada com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/AuthenticatedUser"
                },
                "example" : {
                  "id" : "5ff610e2f6727779c6e6e8b8",
                  "username" : "user1",
                  "token" : "Bearer eyJhbGciOiJIUzUxMiJ9.eyJzdWIiOiJ1c2VyMSIsImlhdCI6MTYxMDA0NDUxOSwiZXhwIjoxNjEyNjM2NTE5fQ.5aGFRATklh3KAyWw_5Y_U4eWsZscU0GlxghpbZobX4kBqOwL8xsA_iiPp0RD_V2wXpqN7jJfz_SL-p_ljs5VoQ",
                  "expiration" : "2021-02-06T15:35:19.799891406-03:00"
                }
              }
            }
          },
          "400" : {
            "description" : "Ao menos um campo da credencial não foi informado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T18:39:58.804116714+00:00",
                  "status" : 400,
                  "error" : "Bad Request",
                  "message" : "Method Argument Not Valid",
                  "path" : "/todo/login",
                  "errors" : [ {
                    "field" : "password",
                    "message" : "must not be blank",
                    "value" : null
                  } ]
                }
              }
            }
          },
          "401" : {
            "description" : "Credencial informada não autorizada",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T18:40:56.292179784+00:00",
                  "status" : 401,
                  "error" : "Unauthorized",
                  "message" : "Invalid username and/or password",
                  "path" : "/todo/login"
                }
              }
            }
          }
        }
      }
    },
    "/logout" : {
      "post" : {
        "tags" : [ "Authentication API" ],
        "summary" : "Remove a autenticação do usuário no sistema",
        "operationId" : "logout",
        "parameters" : [ {
          "in" : "query",
          "name" : "username",
          "schema" : {
            "type" : "string"
          },
          "example" : "lucas",
          "required" : true,
          "description" : "Nome de usuário"
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "type" : "object"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "Operação realizada com sucesso",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object"
                },
                "example" : { }
              }
            }
          },
          "400" : {
            "description" : "Nome de usuário não informado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T19:29:34.210238119+00:00",
                  "status" : 400,
                  "error" : "Bad Request",
                  "message" : "Required String parameter 'username' is not present",
                  "path" : "/todo/logout"
                }
              }
            }
          },
          "401" : {
            "$ref" : "#/components/responses/Unauthorized"
          },
          "403" : {
            "$ref" : "#/components/responses/Forbidden"
          },
          "404" : {
            "description" : "Usuário com nome de usuário informado não encontrado",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiError"
                },
                "example" : {
                  "timestamp" : "2021-01-07T19:51:29.700252263+00:00",
                  "status" : 404,
                  "error" : "Not Found",
                  "message" : "Username not found",
                  "path" : "/todo/logout"
                }
              }
            }
          }
        },
        "security" : [ {
          "todoAuth" : [ ]
        } ]
      }
    },
...
  },
  "components" : {
    "responses" : {
      "Unauthorized" : {
        "description" : "Usuário não autenticado",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            },
            "example" : {
              "timestamp" : "2020-12-31T17:31:04.518+00:00",
              "status" : 401,
              "error" : "Unauthorized",
              "message" : "Invalid token",
              "path" : "..."
            }
          }
        }
      },
      "Forbidden" : {
        "description" : "Usuário não possui permissão para acessar este recurso ou serviço",
        "content" : {
          "application/json" : {
            "schema" : {
              "$ref" : "#/components/schemas/ApiError"
            },
            "example" : {
              "timestamp" : "2021-01-11T21:12:27.253+00:00",
              "status" : 403,
              "error" : "Forbidden",
              "message" : "",
              "path" : "..."
            }
          }
        }
      }
    },
    "schemas" : {
...
      "AuthenticatedUser" : {
        "required" : [ "id", "username", "token", "expiration" ],
        "type" : "object",
        "properties" : {
          "id" : {
            "description" : "Identificador único do usuário",
            "type" : "string"
          },
          "username" : {
            "description" : "Nome de usuário",
            "type" : "string"
          },
          "token" : {
            "description" : "Chave de autenticação do usuário que deve ser enviada no cabeçalho da requisição (Authorization) a cada requisição de acesso à rotas que necessitem de autenticação",
            "type" : "string"
          },
          "expiration" : {
            "description" : "Data de expiração da chave de autenticação",
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "UserCredentials" : {
        "required" : [ "password", "username" ],
        "type" : "object",
        "properties" : {
          "username" : {
            "description" : "Nome de usuário",
            "type" : "string"
          },
          "password" : {
            "description" : "Senha",
            "type" : "string",
            "writeOnly" : true
          }
        }
      },
...
      "ApiError" : {
        "required" : [ "timestamp", "status", "error", "message", "path" ],
        "type" : "object",
        "properties" : {
          "timestamp" : {
            "description" : "Momento em que ocorreu o erro",
            "type" : "string"
          },
          "status" : {
            "description" : "Status HTTP do erro",
            "type" : "integer"
          },
          "error" : {
            "description" : "Mensagem de erro do status HTTP",
            "type" : "string"
          },
          "message" : {
            "description" : "Mensagem de erro lançada pela exceção, ou mensagem personalizada",
            "type" : "string"
          },
          "path" : {
            "description" : "Caminho relativo do endpoint da solicitação em que ocorreu o erro",
            "type" : "string"
          },
          "errors" : {
            "description" : "Lista de erros",
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/Error"
            }
          }
        }
      },
      "Error" : {
        "required" : [ "field", "message", "value" ],
        "type" : "object",
        "properties" : {
          "field" : {
            "description" : "Nome do campo em que ocorreu o erro",
            "type" : "string"
          },
          "message" : {
            "description" : "Mensagem de erro",
            "type" : "string"
          },
          "value" : {
            "description" : "Valor informado que gerou o erro",
            "type" : "string"
          }
        }
      }
    },
    "securitySchemes" : {
      "todoAuth" : {
        "type" : "http",
        "scheme" : "bearer",
        "bearerFormat" : "JWT"
      }
    }
  }
}